+++ src/database/reciept_payment_types/reciept_payment_types.model.js
@@ -5,7 +5,7 @@ export const ReceiptPaymentTypeSchema = {
 		receipt_payment_type_id:  { type: 'string', optional: true },
         receipt_id:  { type: 'string', optional: true },
         payment_type_id: { type: 'int', optional: true },
-        amount: { type: 'int', optional: true },
+        amount: { type: 'double', optional: true },
         syncAction: { type: 'string', optional: true },
         created_at: { type: 'date', optional: true },
         updated_at: { type: 'date', optional: true },
